pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "db-migration"
version = "1.0.0"
description = "Oracle to PostgreSQL migration tool with intelligent agents and a Streamlit chatbot UI."
readme = "README.md"
requires-python = ">=3.8"
authors = [
  { name = "DBMigration Contributors" }
]
dependencies = [
  "cx_Oracle==8.3.0",
  "psycopg2-binary==2.9.9",
  "python-dotenv==1.0.0",
  "pyyaml==6.0.1",
  "tqdm==4.66.1",
  "tabulate==0.9.0",
  "openai>=1.0.0",
  "anthropic>=0.18.0",
  "streamlit>=1.36.0"
]

[project.optional-dependencies]
dev = [
  "pytest",
]

[project.scripts]
db-migrate = "src.migrate:main"


